+++ index.d.ts
@@ -1,7 +1,0 @@
-import { DataTypeAbstract, ModelAttributeColumnOptions } from "sequelize";
-
-declare global {
-  type SequelizeAttributes<T extends { [key: string]: any }> = {
-    [P in keyof T]: string | DataTypeAbstract | ModelAttributeColumnOptions;
-  };
-}