+++ src/service/index.ts
@@ -3,6 +3,7 @@ import BaseServiceSequelize from "./base-service-sequelize";
 import BaseRabbitOperator from "./base-events-rabbitmq";
 import BaseAxiosService from "./base-http-axios";
 import BaseServiceToken from "./base-service-token";
+import BaseCronJobService, { ICronJob, IServiceJob } from "./base-cronjobs";
 
 export {
   BaseServiceSequelize,
@@ -10,4 +11,8 @@ export {
   BaseRabbitOperator,
   BaseAxiosService,
   BaseServiceToken,
+  //jobs
+  BaseCronJobService,
+  ICronJob,
+  IServiceJob,
 };